local_setup/requirements.txt
python-dotenv==1.0.0
fastapi==0.108.0
redis==5.0.1
twilio==8.9.0
uvicorn[standard]==0.22.0